Real_Time_Stock_Pipeline/requirements.txt
kafka-python
pandas
psycopg2-binary
sqlalchemy
streamlit
requests
pyspark